// frontend/src/generated/api.ts
export interface components {
    schemas: {
// ...
        ConnectionsContract: {
            /** Configured */
            configured: boolean;
            /** Picker Configured */
            picker_configured: boolean;
            /** Callback Url */
            callback_url: string;
            /** Allowed Tools */
            allowed_tools: string[];
            /** Connections */
            connections: components["schemas"]["ConnectionContract"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}